✨ Add : search component 분리 및 memo 사용 - 성능 최적화
